chore: Update PropTypes rule in .eslintrc.cjs and remove prop-types import
Update the PropTypes rule in the .eslintrc.cjs file to turn it off. Also, remove the unnecessary prop-types import in the _id.jsx file. This change improves the code consistency and removes unused dependencies.
