+++ .eslintrc.cjs
@@ -26,7 +26,8 @@ module.exports = {
   rules: {
     'react/jsx-no-target-blank': 'off',
     'react-refresh/only-export-components': ['warn', { allowConstantExport: true }],
-
+    'react/prop-types': 'off',
+    
     'no-restricted-imports': [
       'error',
       {
