+++ src/components/AppBar/AppBar.jsx
@@ -14,7 +14,7 @@ import Tooltip from '@mui/material/Tooltip'
 import Typography from '@mui/material/Typography'
 import { useState } from 'react'
 import { ReactComponent as TrelloIcon } from '~/assets/trello.svg'
-import ModeSelect from '~/components/ModeSelect'
+import ModeSelect from '~/components/ModeSelect/ModeSelect'
 import Profile from './Menu/Profile'
 import Recent from './Menu/Recent'
 import Startted from './Menu/Starred'
