+++ src/pages/Boards/BoardBar/BoardBar.jsx
@@ -22,7 +22,6 @@ function BoardBar() {
         gap: 2,
         paddingX: 2,
         overflowX: 'auto',
-        borderBottom: '1px solid #e0e0e0',
         backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#34495e' : '#1976d2')
       }}
     >
