+++ src/pages/Boards/BoardContent/ListColumns/Columns/Columns.jsx
@@ -1,0 +1,143 @@
+import AddCardIcon from '@mui/icons-material/AddCard'
+import Cloud from '@mui/icons-material/Cloud'
+import ContentCopyIcon from '@mui/icons-material/ContentCopy'
+import ContentCut from '@mui/icons-material/ContentCut'
+import ContentPasteIcon from '@mui/icons-material/ContentPaste'
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
+import DragHandleIcon from '@mui/icons-material/DragHandle'
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import Box from '@mui/material/Box'
+import Button from '@mui/material/Button'
+import Divider from '@mui/material/Divider'
+import ListItemIcon from '@mui/material/ListItemIcon'
+import ListItemText from '@mui/material/ListItemText'
+import Menu from '@mui/material/Menu'
+import MenuItem from '@mui/material/MenuItem'
+import Tooltip from '@mui/material/Tooltip'
+import Typography from '@mui/material/Typography'
+import { useState } from 'react'
+import ListCards from './ListCards/ListCards'
+
+export default function Columns() {
+  const [anchorEl, setAnchorEl] = useState(null)
+  const open = Boolean(anchorEl)
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
+  return (
+    <Box
+      sx={{
+        minWidth: '300px',
+        maxWidth: '300px',
+        backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
+        ml: 2,
+        borderRadius: '6px',
+        height: 'fit-content',
+        maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
+      }}
+    >
+      {/* Box Column Header */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnHeaderHeight,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between',
+          p: 2
+        }}
+      >
+        <Typography
+          variant='h6'
+          sx={{
+            fontWeight: 'bold',
+            fontSize: '1rem',
+            cursor: 'pointer'
+          }}
+        >
+          Column Title
+        </Typography>
+        <Box>
+          <Tooltip title='More options'>
+            <ExpandMoreIcon
+              sx={{ color: 'text.primary' }}
+              id='basic-column-dropdown'
+              aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
+              aria-haspopup='true'
+              aria-expanded={open ? 'true' : undefined}
+              onClick={handleClick}
+            />
+          </Tooltip>
+
+          <Menu
+            id='basic-menu-column-dropdown'
+            anchorEl={anchorEl}
+            open={open}
+            onClose={handleClose}
+            MenuListProps={{
+              'aria-labelledby': 'basic-column-dropdown'
+            }}
+          >
+            <MenuItem>
+              <ListItemIcon>
+                <AddCardIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Add new card</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCut fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Cut</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCopyIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Copy</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentPasteIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Paste</ListItemText>
+            </MenuItem>
+
+            <Divider />
+            <MenuItem>
+              <ListItemIcon>
+                <Cloud fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Remove this column</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <DeleteForeverIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Archive this column</ListItemText>
+            </MenuItem>
+          </Menu>
+        </Box>
+      </Box>
+      {/* Box List Card */}
+      <ListCards />
+      {/* Box Column Footer */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnFooterHeight,
+          p: 2,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between'
+        }}
+      >
+        <Button startIcon={<AddCardIcon />}>Add new card</Button>
+        <Tooltip title='Drag to move'>
+          <DragHandleIcon sx={{ cursor: 'pointer' }} />
+        </Tooltip>
+      </Box>
+    </Box>
+  )
+}
