+++ src/pages/Boards/BoardContent/ListColumns/Columns/ListCards/Card/Card.jsx
@@ -1,0 +1,61 @@
+import AttachmentIcon from '@mui/icons-material/Attachment'
+import CommentIcon from '@mui/icons-material/Comment'
+import GroupIcon from '@mui/icons-material/Group'
+import { Card as MuiCard } from '@mui/material'
+import Button from '@mui/material/Button'
+import CardActions from '@mui/material/CardActions'
+import CardContent from '@mui/material/CardContent'
+import CardMedia from '@mui/material/CardMedia'
+import Typography from '@mui/material/Typography'
+
+export default function Card({ temporaryHideMedia }) {
+  if (temporaryHideMedia) {
+    return (
+      <MuiCard
+        sx={{
+          cursor: 'pointer',
+          boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
+          overflow: 'unset'
+        }}
+      >
+        <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
+          <Typography>Card Test 01</Typography>
+        </CardContent>
+      </MuiCard>
+    )
+  }
+
+  return (
+    <MuiCard
+      sx={{
+        cursor: 'pointer',
+        boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
+        overflow: 'unset'
+      }}
+    >
+      <CardMedia
+        sx={{ height: 140 }}
+        image='https://images.unsplash.com/photo-1502086223501-7ea6ecd79368?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxleHBsb3JlLWZlZWR8MTJ8fHxlbnwwfHx8fHw%3D'
+        title='green iguana'
+      />
+      <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
+        <Typography>Viet Doan MERN STACK</Typography>
+      </CardContent>
+      <CardActions
+        sx={{
+          p: '0 4px 8px 4px'
+        }}
+      >
+        <Button size='small' startIcon={<GroupIcon />}>
+          20
+        </Button>
+        <Button size='small' startIcon={<CommentIcon />}>
+          20
+        </Button>
+        <Button size='small' startIcon={<AttachmentIcon />}>
+          20
+        </Button>
+      </CardActions>
+    </MuiCard>
+  )
+}
