+++ src/pages/Boards/BoardContent/ListColumns/ListColumns.jsx
@@ -1,0 +1,51 @@
+import NoteAddIcon from '@mui/icons-material/NoteAdd'
+import Box from '@mui/material/Box'
+import Button from '@mui/material/Button'
+import Columns from './Columns/Columns'
+
+export default function ListColumns() {
+  return (
+    <Box
+      sx={{
+        backgroundColor: 'inherit',
+        width: '100%',
+        height: '100%',
+        display: 'flex',
+        overflowX: 'auto',
+        overflowY: 'hidden',
+        '&::-webkit-scrollbar-track': {
+          m: 2
+        }
+      }}
+    >
+      <Columns />
+      <Columns />
+      <Columns />
+
+      {/* Add new column */}
+      <Box
+        sx={{
+          maxWidth: '200px',
+          minWidth: '200px',
+          mx: 2,
+          borderRadius: '6px',
+          height: 'fit-content',
+          backgroundColor: '#ffffff3d'
+        }}
+      >
+        <Button
+          sx={{
+            color: 'white',
+            width: '100%',
+            justifyContent: 'flex-start',
+            pl: 2.5,
+            py: 1
+          }}
+          startIcon={<NoteAddIcon />}
+        >
+          Add new column
+        </Button>
+      </Box>
+    </Box>
+  )
+}
