--- src/pages/Boards/BoardContent/index.jsx
@@ -1,525 +1,0 @@
-import AddCardIcon from '@mui/icons-material/AddCard'
-import AttachmentIcon from '@mui/icons-material/Attachment'
-import Cloud from '@mui/icons-material/Cloud'
-import CommentIcon from '@mui/icons-material/Comment'
-import ContentCopyIcon from '@mui/icons-material/ContentCopy'
-import ContentCut from '@mui/icons-material/ContentCut'
-import ContentPasteIcon from '@mui/icons-material/ContentPaste'
-import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
-import DragHandleIcon from '@mui/icons-material/DragHandle'
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import GroupIcon from '@mui/icons-material/Group'
-import Box from '@mui/material/Box'
-import Button from '@mui/material/Button'
-import Card from '@mui/material/Card'
-import CardActions from '@mui/material/CardActions'
-import CardContent from '@mui/material/CardContent'
-import CardMedia from '@mui/material/CardMedia'
-import Divider from '@mui/material/Divider'
-import ListItemIcon from '@mui/material/ListItemIcon'
-import ListItemText from '@mui/material/ListItemText'
-import Menu from '@mui/material/Menu'
-import MenuItem from '@mui/material/MenuItem'
-import Tooltip from '@mui/material/Tooltip'
-import Typography from '@mui/material/Typography'
-import { useState } from 'react'
-const COLUMN_HEARDER_HEIGHT = '50px'
-const COLUMN_FOOTER_HEIGHT = '56px'
-
-function BoardContent() {
-  const [anchorEl, setAnchorEl] = useState(null)
-  const open = Boolean(anchorEl)
-  const handleClick = (event) => {
-    setAnchorEl(event.currentTarget)
-  }
-  const handleClose = () => {
-    setAnchorEl(null)
-  }
-  return (
-    <Box
-      sx={{
-        width: '100%',
-        height: (theme) => theme.trello.boardContentHeight,
-        backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#34495e' : '#1976d2'),
-        p: '10px 0'
-      }}
-    >
-      <Box
-        sx={{
-          backgroundColor: 'inherit',
-          width: '100%',
-          height: '100%',
-          display: 'flex',
-          overflowX: 'auto',
-          overflowY: 'hidden',
-          '&::-webkit-scrollbar-track': {
-            m: 2
-          }
-        }}
-      >
-        {/* Box Column 01 */}
-        <Box
-          sx={{
-            minWidth: '300px',
-            maxWidth: '300px',
-            backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
-            ml: 2,
-            borderRadius: '6px',
-            height: 'fit-content',
-            maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
-          }}
-        >
-          {/* Box Column Header */}
-          <Box
-            sx={{
-              height: COLUMN_HEARDER_HEIGHT,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between',
-              p: 2
-            }}
-          >
-            <Typography
-              variant='h6'
-              sx={{
-                fontWeight: 'bold',
-                fontSize: '1rem',
-                cursor: 'pointer'
-              }}
-            >
-              Column Title
-            </Typography>
-            <Box>
-              <Tooltip title='More options'>
-                <ExpandMoreIcon
-                  sx={{ color: 'text.primary' }}
-                  id='basic-column-dropdown'
-                  aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
-                  aria-haspopup='true'
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-
-              <Menu
-                id='basic-menu-column-dropdown'
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  'aria-labelledby': 'basic-column-dropdown'
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopyIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPasteIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                </MenuItem>
-
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Remove this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-          {/* Box List Card */}
-          <Box
-            sx={{
-              p: '0 5px',
-              m: '0 5px',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-              overflowX: 'hidden',
-              overflowY: 'auto',
-              maxHeight: (theme) =>
-                `calc(${theme.trello.boardContentHeight} - ${COLUMN_HEARDER_HEIGHT} - ${COLUMN_FOOTER_HEIGHT} - ${theme.spacing(5)})`,
-              '&::-webkit-scrollbar-thumb': {
-                backgroundColor: '#ced0da'
-              },
-              '&::-webkit-scrollbar-thumb:hover': {
-                backgroundColor: 'bfc2cf'
-              }
-            }}
-          >
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image='https://images.unsplash.com/photo-1502086223501-7ea6ecd79368?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxleHBsb3JlLWZlZWR8MTJ8fHxlbnwwfHx8fHw%3D'
-                title='green iguana'
-              />
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Viet Doan MERN STACK</Typography>
-              </CardContent>
-              <CardActions
-                sx={{
-                  p: '0 4px 8px 4px'
-                }}
-              >
-                <Button size='small' startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<CommentIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<AttachmentIcon />}>
-                  20
-                </Button>
-              </CardActions>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-          {/* Box Column Footer */}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title='Drag to move'>
-              <DragHandleIcon sx={{ cursor: 'pointer' }} />
-            </Tooltip>
-          </Box>
-        </Box>
-
-        {/* Box Column 02 */}
-        <Box
-          sx={{
-            minWidth: '300px',
-            maxWidth: '300px',
-            backgroundColor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
-            ml: 2,
-            borderRadius: '6px',
-            height: 'fit-content',
-            maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
-          }}
-        >
-          {/* Box Column Header */}
-          <Box
-            sx={{
-              height: COLUMN_HEARDER_HEIGHT,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between',
-              p: 2
-            }}
-          >
-            <Typography
-              variant='h6'
-              sx={{
-                fontWeight: 'bold',
-                fontSize: '1rem',
-                cursor: 'pointer'
-              }}
-            >
-              Column Title
-            </Typography>
-            <Box>
-              <Tooltip title='More options'>
-                <ExpandMoreIcon
-                  sx={{ color: 'text.primary' }}
-                  id='basic-column-dropdown'
-                  aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
-                  aria-haspopup='true'
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-
-              <Menu
-                id='basic-menu-column-dropdown'
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  'aria-labelledby': 'basic-column-dropdown'
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopyIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPasteIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                </MenuItem>
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Remove this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-          {/* Box List Card */}
-          <Box
-            sx={{
-              p: '0 5px',
-              m: '0 5px',
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-              overflowX: 'hidden',
-              overflowY: 'auto',
-              maxHeight: (theme) =>
-                `calc(${theme.trello.boardContentHeight} - ${COLUMN_HEARDER_HEIGHT} - ${COLUMN_FOOTER_HEIGHT} - ${theme.spacing(5)})`,
-              '&::-webkit-scrollbar-thumb': {
-                backgroundColor: '#ced0da'
-              },
-              '&::-webkit-scrollbar-thumb:hover': {
-                backgroundColor: 'bfc2cf'
-              }
-            }}
-          >
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image='https://images.unsplash.com/photo-1502086223501-7ea6ecd79368?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxleHBsb3JlLWZlZWR8MTJ8fHxlbnwwfHx8fHw%3D'
-                title='green iguana'
-              />
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Viet Doan MERN STACK</Typography>
-              </CardContent>
-              <CardActions
-                sx={{
-                  p: '0 4px 8px 4px'
-                }}
-              >
-                <Button size='small' startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<CommentIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<AttachmentIcon />}>
-                  20
-                </Button>
-              </CardActions>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rfba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  '&:last-child': { p: 1.5 },
-                  overflow: 'unset'
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-          {/* Box Column Footer */}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title='Drag to move'>
-              <DragHandleIcon sx={{ cursor: 'pointer' }} />
-            </Tooltip>
-          </Box>
-        </Box>
-      </Box>
-    </Box>
-  )
-}
-
-export default BoardContent
